More rebust testing because new metrics and exchanges could happen
Basically, the testing was just too specific. The approach here was
to describe the tests in a more general fashion instead of looking
for specific values as output. Any deviation from the expected value
caused a FAIL even when the API was working as intended.

No version bump will be made since the dist provided isn't changed.